requirements.txt
feedparser
pymsteams
python-dotenv
